requirements.txt
alembic==1.4.2
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
regex==2020.6.8
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
